Move getStateAnalystsEmails and test to its own directory.
